Support Develocity and Gradle Enterprise remote build cache connectors
Adds a new command line argument for Gradle experiment 5, `-y` or
`--remote-build-cache-type`, allowing the user to specify a remote build
cache type to use. When specified, valid values are: `develocity`,
`gradle-enterprise`, or `http`. When the argument is specified, it will
always take precedence over what is configured in the build.

If the argument is not specified, the remote build cache type of the
first build will be fetched from the Develocity API and used for the
second build.

If the argument is not specified and the user does not have access to
the Develocity API, then the existing remote build cache configuration
will be used.